feat(upgrade): restructure upgrade system by removing old files and adding new modular components for better maintainability fix(index.html): update script source path to reflect new directory structure feat(counter): add counter functionality to track score and update display feat(api): implement loadUpgrades function to fetch upgrade data from JSON file feat(upgrade): create new upgrade card, content, icon, and cps components for improved UI rendering
