--- assets/js/shop/upgrade/stat/stat.js
@@ -1,12 +1,0 @@
-import {createElement} from "../upgrade";
-
-export function createUpgradeStats(cps) {
-    const stats = createElement("div", ["upgradeStats", "flex"]);
-    const cpsElement = createElement("p");
-
-    cpsElement.innerHTML = `<span class="cps">+${cps}</span><br>par sec`;
-
-    stats.appendChild(cpsElement);
-
-    return stats;
-}
