+++ index.html
@@ -12,7 +12,7 @@
     <link href="https://cdn.jsdelivr.net/npm/remixicon/fonts/remixicon.css"
           rel="stylesheet"
     />
-    <script type="module" src="/assets/js/main.js"></script>
+    <script type="module" src="/js/main.js"></script>
 </head>
 <body>
 <main class="flex">
