+++ js/main.js
@@ -1,7 +1,6 @@
 import { loadUpgrades } from "./shop/api.js";
 import { showUpgrades } from "./shop/upgrade/upgrade.js";
-import {incrementScore, updateScoreDisplay} from "./counter/counter";
-
+import {incrementScore, updateScoreDisplay} from "./counter/counter.js";
 
 const cookie = document.getElementById("cookieButton");
 const score = document.getElementById("score");
