+++ js/shop/api.js
@@ -1,6 +1,5 @@
 export async function loadUpgrades() {
-    const response = await fetch("../../../shop.json");
-
+    const response = await fetch("../shop.json");
     if (!response.ok) {
         throw new Error("Failed to load upgrade");
     }
