+++ js/shop/upgrade/icon/icon.js
@@ -1,0 +1,12 @@
+import {createElement} from "../upgrade.js";
+
+export function createUpgradeIcon(iconClass) {
+    const iconContainer = createElement("div", ["flex"]);
+    const icon = createElement("i");
+
+    icon.className = iconClass;
+
+    iconContainer.appendChild(icon);
+
+    return iconContainer;
+}
