+++ package.json
@@ -3,5 +3,8 @@
     "@eslint/js": "^10.0.1",
     "eslint": "^10.3.0",
     "globals": "^17.6.0"
+  },
+  "scripts": {
+    "lint": "eslint js/*.js"
   }
 }
